requirements.txt
flask==2.0.1
gunicorn==20.1.0
pymongo==4.3.3
python-dotenv==0.19.0
werkzeug==2.0.3
tqdm
pandas
